requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.1
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
orjson==3.9.10
passlib==1.7.4
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
wheel==0.37.1
